Refactoring; shifted some responsibilities from sysent to vfs
