--- include/basic_dev.h
@@ -1,7 +1,0 @@
-#ifndef __BASIC_DEV_H__
-#define __BASIC_DEV_H__
-#include <file.h>
-
-int dev_null_open(file_t *f, int flags, int mode);
-
-#endif /* __BASIC_DEV_H__ */
